+++ src/components/copy-button/copy-button.ts
@@ -207,8 +207,8 @@ export default class WaCopyButton extends WebAwesomeElement {
         <wa-tooltip
           class=${classMap({
             'copy-button': true,
-            'copy-button--success': this.status === 'success',
-            'copy-button--error': this.status === 'error',
+            'copy-button-success': this.status === 'success',
+            'copy-button-error': this.status === 'error',
           })}
           for="copy-button"
           placement=${this.tooltipPlacement}